requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gymnasium>=0.29.1
pydantic>=2.7.0
openai>=1.40.0
openenv-core>=0.2.0
python-dotenv>=1.0.0
